+++ Plugins/ActorPools/Source/NexusActorPools/Private/NActorPoolListViewEntry.cpp
@@ -56,8 +56,8 @@ void UNActorPoolListViewEntry::Refresh() const
 		return;
 	}
 
-	const int32 InCount = Pool->GetInCount();
-	const int32 OutCount = Pool->GetOutCount();
+	const int32 InCount = Pool->GetOutCount();
+	const int32 OutCount = Pool->GetSpawnedCount();
 	const int32 Total = InCount + OutCount;
 	
 	LeftText->SetText(FText::AsNumber(InCount));