+++ Plugins/ActorPools/Source/NexusActorPoolsEditor/Private/NActorPoolsEditorCommands.cpp
@@ -70,10 +70,10 @@ void FNActorPoolsEditorCommands::AddActorPoolMethods(const UObject* EditedObject
 		FBlueprintEditor* BlueprintEditor = static_cast<FBlueprintEditor*>(EditorInstance);
 		UBlueprint* Blueprint = BlueprintEditor->GetBlueprintObj();
 		
-		N_ACTOR_POOL_INVOKE_METHODS(OnCreated, NEXUS::ActorPools::InvokeMethods::OnCreated, NEXUS::ActorPools::InvokeMethods::Category)
-		N_ACTOR_POOL_INVOKE_METHODS(OnSpawned, NEXUS::ActorPools::InvokeMethods::OnSpawned, NEXUS::ActorPools::InvokeMethods::Category)
-		N_ACTOR_POOL_INVOKE_METHODS(OnReturn, NEXUS::ActorPools::InvokeMethods::OnReturn, NEXUS::ActorPools::InvokeMethods::Category)
-		N_ACTOR_POOL_INVOKE_METHODS(OnDestroyed, NEXUS::ActorPools::InvokeMethods::OnDestroyed, NEXUS::ActorPools::InvokeMethods::Category)
+		N_ACTOR_POOL_INVOKE_METHODS(OnCreated, NEXUS::ActorPools::InvokeMethods::OnCreatedByActorPool, NEXUS::ActorPools::InvokeMethods::Category)
+		N_ACTOR_POOL_INVOKE_METHODS(OnSpawned, NEXUS::ActorPools::InvokeMethods::OnSpawnedFromActorPool, NEXUS::ActorPools::InvokeMethods::Category)
+		N_ACTOR_POOL_INVOKE_METHODS(OnReturn, NEXUS::ActorPools::InvokeMethods::OnReturnToActorPool, NEXUS::ActorPools::InvokeMethods::Category)
+		N_ACTOR_POOL_INVOKE_METHODS(OnDestroyed, NEXUS::ActorPools::InvokeMethods::OnDestroyedByActorPool, NEXUS::ActorPools::InvokeMethods::Category)
 		
 		FBlueprintEditorUtils::MarkBlueprintAsStructurallyModified(Blueprint);
 	}