+++ Plugins/ActorPools/Source/NexusActorPoolsEditor/Tests/INActorPoolItemTests.cpp
@@ -39,15 +39,15 @@ N_TEST(INActorPoolItemTests_Events,
 
 		CHECK_EQUALS("Check return to pool state event occurred.", TestObject->State, static_cast<uint8>(ENActorOperationalState::Disabled))
 		CHECK_EQUALS("Delegate should have fired exactly once after return.", TestObject->Counter, 1)
-		CHECK_EQUALS("Pool should have 1 actor in after return.", Pool.GetInCount(), 1)
-		CHECK_EQUALS("Pool should have 0 actors out after return.", Pool.GetOutCount(), 0)
+		CHECK_EQUALS("Pool should have 1 actor in after return.", Pool.GetAvailableCount(), 1)
+		CHECK_EQUALS("Pool should have 0 actors out after return.", Pool.GetSpawnedCount(), 0)
 
 		Pool.Spawn(FVector::Zero(), FRotator::ZeroRotator);
 
 		CHECK_EQUALS("Check spawn state event occurred.", TestObject->State, static_cast<uint8>(ENActorOperationalState::Enabled))
 		CHECK_EQUALS("Delegate should have fired exactly twice after spawn.", TestObject->Counter, 2)
-		CHECK_EQUALS("Pool should have 0 actors in after spawn.", Pool.GetInCount(), 0)
-		CHECK_EQUALS("Pool should have 1 actor out after spawn.", Pool.GetOutCount(), 1)
+		CHECK_EQUALS("Pool should have 0 actors in after spawn.", Pool.GetAvailableCount(), 0)
+		CHECK_EQUALS("Pool should have 1 actor out after spawn.", Pool.GetSpawnedCount(), 1)
 
 		Pool.Clear();